Adds helper function to check if an orderby clause is sortable
The way that WP and this library handle specifying orderby fields is not working well. The orderby clause from WordPress is the actual meta_key that posts should be sorted by, but PostType expects the orderby clause to be the column_key. This works fine until the meta_key and the column_key are different.

This commit makes the check more extensive and supports the actual sort behavior as implemented by WordPress.